update sql query to convert id to int to ensure type of return value

+++ lib/username_completion.rb
@@ -21,7 +21,7 @@ class UsernameCompletion
   def query
     <<-SQL
       select
-        id,
+        id::int,
         login,
         display_name
